requirements.txt
#server
fastapi==0.88.0
uvicorn==0.20.0
gunicorn==20.1.0
#openai
openai==0.25.0
#telegram
python-telegram-bot==13.15
#env
python-dotenv==0.21.0